feat: 연락처 서치바 기능 구현, extension 파일 분리
