+++ contact-management/contact-management/Controller/ContactListViewController+TableView.swift
@@ -1,0 +1,37 @@
+//
+//  ContactListViewController+TableView.swift
+//  contact-management
+//
+//  Created by Roh on 1/19/24.
+//
+
+import UIKit
+
+extension ContactListViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return filteredDataSource.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "myCell", for: indexPath)
+        
+        let item = filteredDataSource[indexPath.row]
+        cell.textLabel?.text = "\(item.name) (\(item.age))"
+        cell.detailTextLabel?.text = item.phoneNumber
+        return cell
+    }
+    
+    func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
+        if editingStyle == .delete {
+            contactListStorage?.deleteContact(indexPath.row)
+            tableView.deleteRows(at: [indexPath], with: .fade)
+        }
+    }
+}
+
+extension ContactListViewController: UITableViewDelegate {
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        tableView.deselectRow(at: indexPath, animated: true)
+    }
+}
+
